requirements.txt
# GramSetu - Python Dependencies
# Python 3.11+

# ============================================
# Core Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ============================================
# AWS Services (Member 2 & 3)
# ============================================
boto3==1.34.34
botocore==1.34.34
aioboto3==12.3.0

# ============================================
# LLM & AI
# ============================================
# LangChain for orchestration
langchain==0.1.4
langchain-aws==0.1.0
langchain-community==0.0.16

# OpenAI (fallback)
openai==1.10.0

# Anthropic (for Claude)
anthropic==0.8.1

# ============================================
# Browser Automation (Member 2)
# ============================================
playwright==1.41.0
selenium==4.17.2
pyautogui==0.9.54
pillow==10.2.0

# ============================================
# Audio Processing (Member 1)
# ============================================
# Audio I/O
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# Noise suppression
noisereduce==3.0.0
webrtcvad==2.0.10

# Silero VAD
torch==2.1.2
torchaudio==2.1.2

# ============================================
# Computer Vision & OCR (Member 3)
# ============================================
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
pytesseract==0.3.10

# Document processing
pdf2image==1.17.0
pypdf2==3.0.1

# PII detection
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351

# ============================================
# Database (Member 4)
# ============================================
# PostgreSQL
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1
celery==5.3.6

# ============================================
# HTTP & API Clients
# ============================================
httpx==0.26.0
aiohttp==3.9.3
requests==2.31.0

# ============================================
# WhatsApp Integration (Member 4)
# ============================================
twilio==8.11.1
python-telegram-bot==20.7

# ============================================
# Utilities
# ============================================
# Phonetic matching
metaphone==0.6
jellyfish==1.0.3

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# JSON/YAML
pyyaml==6.0.1
orjson==3.9.13

# Validation
validators==0.22.0
phonenumbers==8.13.29

# ============================================
# Security & Encryption
# ============================================
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ============================================
# Monitoring & Logging
# ============================================
loguru==0.7.2
sentry-sdk==1.40.0
prometheus-client==0.19.0

# ============================================
# Testing
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.16.0
faker==22.5.1

# ============================================
# Development Tools
# ============================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# ============================================
# Data Processing
# ============================================
pandas==2.2.0
numpy==1.26.3

# ============================================
# Async & Concurrency
# ============================================
asyncio==3.4.3
aiofiles==23.2.1
tenacity==8.2.3

# ============================================
# WebSocket (Real-time updates)
# ============================================
websockets==12.0
python-socketio==5.11.0

# ============================================
# Rate Limiting
# ============================================
slowapi==0.1.9

# ============================================
# Environment-specific
# ============================================
# Install Playwright browsers after pip install
# Run: playwright install chromium
